requirements.txt
openai
numpy
scikit-learn
matplotlib
python-dotenv
tiktoken
qdrant-client
pymongo
google-genai
groq
requests
pandas
aiohttp
beautifulsoup4
asyncio
uvicorn
fastapi
python-multipart
streamlit
tavily-python
tenacity>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiolimiter>=1.1.0